soc: qcom: icc-bwmon: Set default thresholds dynamically
Currently we use predefined threshold values. This works, but does not
really scale well, as it may differ between SoCs due to LPDDR generation
and/or DDR controller setup. All of the data we need for that is already
provided in the device tree, anyway.

Change that to:
* 0 for low (as we've been doing up until now)
* BW_MIN/2 for mid
* BW_MIN for high

The mid value is chosen for a "low enough" bw to nudge bwmon into
slowing down if it starts too high from the bootloader.

The high value corresponds to the upper barrier which - when crossed -
raises an interrupt in the third zone, signaling a need for upping
the bw.

This only changes the values programmed at probe time, as high and med
thresholds are updated at interrupt, based on the OPP table from DT.

Signed-off-by: Konrad Dybcio <konrad.dybcio@linaro.org>