dt-bindings: interconnect: qcom,msm8996: Separate out MNoC_AHB
Sadly, the initial revision of this driver misrepresented the NoC
structure, which calls for an ABI-breaking change that separates out
part of MNoC into a new entity - MNoC_AHB.  That _AHB bus is clocked
from a different source and the aggregate bandwidth on the bus should
be used as an input for setting the rate on the _AHB clock.

Separate it out to enable proper functionality.

Fixes: 3e9fdc6b73ca ("dt-bindings: interconnect: Add Qualcomm MSM8996 DT bindings")
Signed-off-by: Konrad Dybcio <konrad.dybcio@linaro.org>