soc: qcom: Add a driver for CPR3+
Introduce a new driver (based on qcom-cpr for CPRv1 IP) for the newer
Qualcomm Core Power Reduction hardware, known downstream as CPR[34h]
(h for hardened).

In these new CPR versions, support for various new features was introduced.
That includes:
* voltage reduction for the GPU
* security hardening
* a new way of controlling CPU DVFS, based on internal communication
  between CPRh and Operating State Manager MCUs.

The CPR v3, v4 and CPRh are present in a broad range of SoCs, from the
mid-range to the high end ones including, but not limited to:
MSM8953/8996/8998 and SDM630/636/660/845.

Note that to reduce the giant review and testing matrix of the driver, this
patch (admittedly, somewhat confusingly but for good reasons) omits support
for CPR*3* specifically, which is otherwise quite straightforward to add.

Signed-off-by: AngeloGioacchino Del Regno <angelogioacchino.delregno@somainline.org>
[Konrad: rebase, a whole lot of cleanup/fixes]
Signed-off-by: Konrad Dybcio <konrad.dybcio@linaro.org>