arm64: dts: qcom: sc8280xp: Add cpufreq-based bus votes
On RPMh-based SoCs, the memory bus looks roughly like this:

        [B]     [B]
         |       |
[CPU_SS]---[LLC]---[DDR]

Where LLC is Last Level Cache and B is a BWMON.

This poses a serious problem with the current setup, as the BWMONs can't
see things that never leave the previous pipeline stage (e.g. when the
workload is mostly contained in the cache).

Combined with how we currently interpret the data that BWMONs spit out
(instant reaction with 100% trust in current usage, no averaging or
analyzing previous and possible future usage patterns), this results in
esssentially permanent bandwidth starvation.

To (hopefully temporarily) relieve this, introduce CPU<->LLCC and
LLCC<->DDR interconnect votes, bound to CPU frequencies. This lets us
piggyback off of cpufreq going up or down to make a "decent enough" guess
whether we're going to need more or less DRAM (or LLC) bandwidth, at the
cost of increased power usage.

This has been determined to be the least painful solution for the time
being.

Signed-off-by: Konrad Dybcio <konrad.dybcio@linaro.org>