MPM hooking upping
This series hooks up MPM (which can finally be done after winning the
lengthy bindings fight) on 6375, 8996 and 2290 as a good start.

The hardware in question is present on all RPM SMD SoCs (so, roughly
msm8974 and up), which means many more platforms are there to come.

MPM is one of the key pieces for allowing full system power collapse
(and perhaps as importantly, resume) and serves roughly the same role
as PDC on RPMh SoCs for this purpose.

To: Andy Gross <agross@kernel.org>
To: Bjorn Andersson <andersson@kernel.org>
To: Rob Herring <robh+dt@kernel.org>
To: Krzysztof Kozlowski <krzysztof.kozlowski+dt@linaro.org>
To: Conor Dooley <conor+dt@kernel.org>
Cc: Konrad Dybcio <konrad.dybcio@somainline.org>
Cc: linux-arm-msm@vger.kernel.org
Cc: devicetree@vger.kernel.org
Cc: linux-kernel@vger.kernel.org
Signed-off-by: Konrad Dybcio <konrad.dybcio@linaro.org>

---
Changes in v2:
- EDITME: describe what is new in this series revision.
- EDITME: use bulletpoints and terse descriptions.
- Link to v1: https://lore.kernel.org/r/20231215-topic-mpm_dt-v1-0-c6636fc75ce3@linaro.org



--- b4-submit-tracking ---
# This section is used internally by b4 prep for tracking purposes.
{
  "series": {
    "revision": 2,
    "change-id": "20231215-topic-mpm_dt-3811720cf884",
    "prefixes": [],
    "history": {
      "v1": [
        "20231215-topic-mpm_dt-v1-0-c6636fc75ce3@linaro.org"
      ]
    }
  }
}