thermal: core: Print out thermal zone name on update_temperature error
Currently, we can encounter an error like this:

  thermal thermal_zone48: failed to read out thermal zone (-19)

It's good to know that there's been an issue, but on some occasions
(like the error happening in the middle of a platform crash), one may
not be able to look up what kind of thermal zone that is.

Add the TZ name to the error message in order to speed up debugging.

Signed-off-by: Konrad Dybcio <konrad.dybcio@linaro.org>