firmware/psci: Set pm_set_resume/suspend_via_firmware() on qcom
Qualcomm platforms implementing PSCI (ab)use CPU_SUSPEND for entering
various stages of suspend, across the SoC. These range from a simple WFI
to a full-fledged power collapse of the entire chip (mostly, anyway).

Some device drivers are curious to know whether "the firmware" (which is
often assumed to be ACPI) takes care of suspending or resuming the
platform. Set the flag that reports this behavior on the aforementioned
chips.

Signed-off-by: Konrad Dybcio <konrad.dybcio@linaro.org>