arm64: dts: qcom: sc8280xp: Hook up actual CPU cooling
Based on the measurements taken on a ThinkPad X13s (a passively cooled
13" laptop), specify sustainable-power values for each cluster to make
the power_allocator thermal governor accessible. This is not ideal, as
each device is different, but we gotta start somewhere.

Also, expand the definitions of the CPU-adjacent thermal zones to make
sure the CPUs are actually throttled based on the sensors' values (and
not only based on e.g. skin-temp values on the X13s).

Signed-off-by: Konrad Dybcio <konrad.dybcio@linaro.org>