devfreq: Add Qualcomm TZ Adreno governor
The Qualcomm TrustZone is known for being a very broad creation. One of
the things it provides, is a devfreq governor for the onboard Adreno
GPU. This governor is interfaced with through secure calls.

There are two modes of operation:
* context-aware (the scheduler is fed the # of active submits before
  they actually hit the GPU)
* "normal" (it's not fed that data)
with context-aware being the preferred option and a fallback in place
in case it's not implemented.

Introduce a driver to make use of what is believed to be Qualcomm's
secret sauce algorhitms that are probably way more educated about the
hw than the simple_ondemand we currently use.

Signed-off-by: Konrad Dybcio <konrad.dybcio@linaro.org>